image_editor/image_tools/requirements.txt
numpy==1.13.1
scipy==0.19
Pillow
cython
matplotlib
scikit-image
tensorflow==1.3.0
keras==2.0.8
opencv-python
h5py
imgaug==0.2.6
IPython[all]
pandas==0.20
seaborn==0.8
pywavelets==0.5.2